requirements.txt
aiohttp
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.10.0; python_version >= '3.9'
cachetools==5.5.2; python_version >= '3.7'
certifi==2025.8.3; python_version >= '3.7'
cffi==1.17.1; python_version >= '3.8'
charset-normalizer==3.4.3; python_version >= '3.7'
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-speech==2.33.0
google-cloud-vision==3.10.2
google-genai==1.32.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gtts==2.5.1
grpcio==1.74.0; python_version >= '3.9'
grpcio-status==1.71.2; python_version >= '3.9'
h11==0.16.0; python_version >= '3.8'
httpcore==1.0.9; python_version >= '3.8'
httplib2==0.22.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
httpx==0.28.1; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
pillow==11.3.0; python_version >= '3.9'
openai==0.28
fastapi
proto-plus==1.26.1; python_version >= '3.7'
protobuf==5.29.5; python_version >= '3.8'
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.2; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pydantic==2.11.7; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
pyee==13.0.0; python_version >= '3.8'
pyparsing==3.2.3; python_version >= '3.9'
python-ffmpeg==2.0.12; python_version >= '3.7'
python-telegram-bot==22.3; python_version >= '3.9'
requests==2.32.4; python_version >= '3.8'
rsa==4.9.1; python_version >= '3.6' and python_version < '4'
sniffio==1.3.1; python_version >= '3.7'
srt==3.5.3; python_version >= '2.7'
tqdm==4.67.1; python_version >= '3.7'
typing-extensions==4.14.1; python_version >= '3.9'
typing-inspection==0.4.1; python_version >= '3.9'
uritemplate==4.2.0; python_version >= '3.9'
urllib3==2.5.0; python_version >= '3.9'
websockets==15.0.1; python_version >= '3.9'
